reorganizing files and commenting out fdn code
