+++ include/symbol-match/symbol-match.hpp
@@ -1,25 +1,23 @@
-#pragma once
+// #pragma once
 
-#include "device/device.hpp"
-#include "state/state-machine.hpp"
-#include "symbol-match-states.hpp"
-#include "symbol-match/symbol-manager.hpp"
+// #include "device/device.hpp"
+// #include "state/state-machine.hpp"
+// #include "symbol-match-states.hpp"
+// #include "symbol-match/symbol-manager.hpp"
 
-class SymbolWirelessManager;
+// class SymbolWirelessManager;
 
-class SymbolWirelessManager;
+// constexpr int SYMBOLMATCH_APP_ID = 1;
 
-constexpr int SYMBOLMATCH_APP_ID = 1;
+// class SymbolMatch : public StateMachine {
+// public:
+//     SymbolMatch(Device* FDN, SymbolWirelessManager* symbolWirelessManager);
+//     ~SymbolMatch();
 
-class SymbolMatch : public StateMachine {
-public:
-    SymbolMatch(Device* FDN, SymbolWirelessManager* symbolWirelessManager);
-    ~SymbolMatch();
+//     void populateStateMap() override;
 
-    void populateStateMap() override;
-
-private:
-    RemoteDeviceCoordinator* remoteDeviceCoordinator;
-    SymbolManager* symbolManager;
-    SymbolWirelessManager* symbolWirelessManager;
-};
+// private:
+//     RemoteDeviceCoordinator* remoteDeviceCoordinator;
+//     SymbolManager* symbolManager;
+//     SymbolWirelessManager* symbolWirelessManager;
+// };
