+++ include/wireless/symbol-wireless-manager.hpp
@@ -6,7 +6,7 @@
 
 #include "device/drivers/serial-wrapper.hpp"
 #include "device/wireless-manager.hpp"
-#include "symbol-match/symbol.hpp"
+#include "symbol.hpp"
 
 class RemoteDeviceCoordinator;
