+++ src/symbol-match/symbol-manager.cpp
@@ -1,59 +1,59 @@
-#include "symbol-match/symbol-manager.hpp"
+// #include "symbol-match/symbol-manager.hpp"
 
-SymbolManager::SymbolManager() {
-    symbols[SymbolPosition::LEFT] = new Symbol();
-    symbols[SymbolPosition::RIGHT] = new Symbol();
-}
+// SymbolManager::SymbolManager() {
+//     symbols[SymbolPosition::LEFT] = new Symbol();
+//     symbols[SymbolPosition::RIGHT] = new Symbol();
+// }
 
-SymbolManager::~SymbolManager() {
-    for(auto& symbol : symbols) {
-        delete symbol.second;
-    }
-    symbols.clear();
-}
+// SymbolManager::~SymbolManager() {
+//     for(auto& symbol : symbols) {
+//         delete symbol.second;
+//     }
+//     symbols.clear();
+// }
 
-void SymbolManager::refreshSymbols() {
-    for(auto& symbol : symbols) {
-        symbol.second->setRandomSymbol();
-    }
-}
+// void SymbolManager::refreshSymbols() {
+//     for(auto& symbol : symbols) {
+//         symbol.second->setRandomSymbol();
+//     }
+// }
 
-Symbol* SymbolManager::getSymbol(SymbolPosition position) {
-    return symbols[position];
-}
+// Symbol* SymbolManager::getSymbol(SymbolPosition position) {
+//     return symbols[position];
+// }
 
-const char* SymbolManager::getSymbolGlyph(SymbolPosition position) {
-    return symbols[position]->getSymbolGlyph();
-}
+// const char* SymbolManager::getSymbolGlyph(SymbolPosition position) {
+//     return symbols[position]->getSymbolGlyph();
+// }
 
-void SymbolManager::validateSymbols(const uint8_t* fdnMac) {
-    (void)fdnMac;
-}
+// void SymbolManager::validateSymbols(const uint8_t* fdnMac) {
+//     (void)fdnMac;
+// }
 
-SimpleTimer* SymbolManager::getRefreshTimer() {
-    return &refreshTimer;
-}
+// SimpleTimer* SymbolManager::getRefreshTimer() {
+//     return &refreshTimer;
+// }
 
-int SymbolManager::getTimeLeftToRefresh() {
-    return refreshInterval - refreshTimer.getElapsedTime();
-}
+// int SymbolManager::getTimeLeftToRefresh() {
+//     return refreshInterval - refreshTimer.getElapsedTime();
+// }
 
-void SymbolManager::resetRefreshTimer() {
-    refreshTimer.setTimer(refreshInterval);
-}
+// void SymbolManager::resetRefreshTimer() {
+//     refreshTimer.setTimer(refreshInterval);
+// }
 
-void SymbolManager::setLeftMatched(bool matched) {
-    leftMatched = matched;
-}
+// void SymbolManager::setLeftMatched(bool matched) {
+//     leftMatched = matched;
+// }
 
-void SymbolManager::setRightMatched(bool matched) {
-    rightMatched = matched;
-}
+// void SymbolManager::setRightMatched(bool matched) {
+//     rightMatched = matched;
+// }
 
-bool SymbolManager::isLeftMatched() {
-    return leftMatched;
-}
+// bool SymbolManager::isLeftMatched() {
+//     return leftMatched;
+// }
 
-bool SymbolManager::isRightMatched() {
-    return rightMatched;
-}
+// bool SymbolManager::isRightMatched() {
+//     return rightMatched;
+// }
